requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
SQLAlchemy==2.0.29
alembic==1.13.1
psycopg[binary]==3.1.18
python-dotenv==1.0.1
pydantic==2.7.1